Purchase bill changes (#795)
* Changes to accomodate purchase bill

* PFM-2678 | PFM-2141 | Expense calculator | Integration with expense | Purchase initial Draft

* PFM-2678 | PFM-2141 | Expense calculator | Integration with expense | Purchase initial Draft

* PFM-2678 | PFM-2141 | Expense calculator | Integration with expense | Purchase initial Draft

* PFM-2678 | PFM-2141 | Expense calculator | Integration with expense | Purchase initial Draft

---------

+++ backend/expense-calculator/src/main/java/org/egov/digit/expense/calculator/web/models/Pagination.java
@@ -26,15 +26,15 @@ public class Pagination {
 	@JsonProperty("limit")
 
 	@DecimalMax("100")
-	private Double limit = 10d;
+	private Integer limit = 10;
 
 	@JsonProperty("offSet")
 
-	private Double offSet = 0d;
+	private Integer offSet = 0;
 
 	@JsonProperty("totalCount")
 
-	private Double totalCount = null;
+	private Integer totalCount = null;
 
 	@JsonProperty("sortBy")
